+++ NEWS.md
@@ -1,12 +1,19 @@
 # MetaboAnnotation 1.1
 
-## Changes in 1.1.3
+## Changes in 1.1.4
 
 - `matchSpectra`: support a `CompDb` with parameter `target`.
 - Add `CompAnnotionSource` classes to support definition of references to
   annotation resources.
 - Add `CompDbSource` class defining a reference to a `CompDb` database.
 - `matchSpectra`: support for `CompDbSource` with parameter `target`.
+
+## Changes in 1.1.3
+
+- Extend `filterMatches` framework (issue
+  [#86](https://github.com/rformassspectrometry/MetaboAnnotation/issues/86)).
+  `ScoreThresholdParam` added to perform filtering the matches based on a
+  threshold for the scores.
 
 ## Changes in 1.1.2